net: Provide an alternative to libvirt.is_libvirt_device
The dhclient_monitor has been using libvirt.is_libvirt_device to
determine if a processed interface belongs to an owned network.

As part of the effort to remove the dependency on libvirt and to
correctly check for network interfaces, the networks_northbound_ifaces
is introduced, reporting all the northbound ifaces of all networks.

Change-Id: Ia0b4fc20a65137e5f40ebc624b18b140f6bdd32c
Signed-off-by: Edward Haas <edwardh@redhat.com>