operation: Prepare for integrating CommandStream
operation.Command supports thread-safe abort, but cannot be used for
commands that needs progress, because it is caching entire process
output.

Move the state management code during startup into _start_process() and
extract _finalize() method for the shutdown state management code.  We
will reuse these methods for the version streaming output to the caller
with CommandStream.

After these changes, _wait_for_process() is only a call to
Popen.communicate(), so it was inlined into run().

Change-Id: Iff363e710fcdc5d6cd5843ca2239973c89a69dd3
Signed-off-by: Nir Soffer <nsoffer@redhat.com>