blockVolume: validate initial size with max size
When creating a volume, there is an optional parameter called initial
size used for QCOW volumes.

The initial size was validated against the volume size. If it was
greater than the size, an InvalidParameterException was thrown.

However, there are cases where the 'truesize' can be greater than the
'capacity' due to the qcow2 overhead and alignment to extent size.

In this patch, the validation of the 'initial size' will be against the
'max size' which is computed based on the 'size', the qcow2 overhead and
extent size.

Change-Id: Id417d8f0a299baf6636fd685cbbc7576a80fccef
Bug-Url: https://bugzilla.redhat.com/1434927
Signed-off-by: Fred Rolland <frolland@redhat.com>