virt: metadata: get rid of _XMLWrapper
We added _XMLWrapper out of necessity, because we possibly
need to *replace* (or at very least update) the object
returned by the interal _domain_xml context manager.

In this patch we reduce the amount of cerimony replacing
the class with much simpler plain list wrapper.

Also, we rename the _domain_xml helper to a more correct
_metadata_xml.
No changes in behaviour for the clients of the metadata
module, tests included.

Change-Id: I2e454495b4189e1656f9357e1d26bda105c7ad50
Signed-off-by: Francesco Romani <fromani@redhat.com>