logging: Print volume IDs instead of object references
When logging the volume chain we are currently logging references to volume
objects which is useless.  Log the volume ID for each volume instead.
Also, move the logging to the caller of Image.getChain so we can add a
little more context and omit duplicate messages where they don't add
value.

For example:

INFO  (merge/3c0ba002) [storage.Image]
    sd=7271d11d-e58d-46cc-b8ba-024830612abe
    img=ce510487-53da-4916-8a03-154aa2735e6b
    chain=2938ad34-bb16-4015-bc2f-44f362774708
        < c55b088d-e700-4648-906d-6bd848625e82
        < ac6ed641-eef2-41be-9ffd-f3e0369af0da
        < 66765728-b659-4c47-b25e-5041b5ae165a (top)  (image:251)

Change-Id: Ib82a1aba410f279725bdcf87fde15351f5980f6a
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1450646
Signed-off-by: Adam Litke <alitke@redhat.com>