virt: Don't eject CD forcefully
Changing or ejecting a CDs is performed forcefully, as with using a paperclip
on a physical device.  The reason for doing that is described in
https://bugzilla.redhat.com/626334, which doesn't look like something valid
these days.  Removing a CD forcefully, without notifying the guest OS, is
hardly a proper action and may cause confusion in the guest OS environment,
such as the CD being apparently still mounted and producing I/O errors when
trying to access it.

This patch makes the CD eject cleanly.  Only if that attempt fails (and is
signaled as a failure from libvirt) then the CD is ejected forcefully.  When a
CD is accessed while being ejected, there can still be problems, but at least
modern RHEL guest OS apparently handles the situation properly: The CD is
unmounted, no longer accessible, and another CD can be mounted.

A different situation is with changing the CD.  When something accesses the CD,
such as a process having its current working directory there, weird things may
happen with either forceful or clean CD change.  The change may not be properly
recognized, errors may happen when the CD is accessed, or the CD may not be
visible at all.  Under such circumstances, when there is no clearly superior
option, it's better to stick with the current behavior.  So we keep changing
CDs forceful.

Bug-Url: https://bugzilla.redhat.com/1428514
Backport-To: 4.1
Change-Id: I71067ea1783b1121bc318bfa2a6fec881938295b
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>