operation: Support streaming command output
operation.Command can be used now when the caller need to process
command's stdout for example when reporting progress.

Example usage:

    out = bytearray()
    for data in proc.watch():
        out +=  data
        # handle progress messages in out...

If command failed, Command.watch() will raise the usual cmdutils.Error,
including the collected stderr. Since the output was passed to the
caller, it will not be included in the error.

This required before utils.CommandStream, which is harder to use since
you need to define callbacks for receiving stdout and stderr data,
collect stderr yourself, and reimplement this in every command that
needs progress.

Change-Id: I976314ce3e9a1b4ba77d74c3c1ad742cc0e9fe80
Signed-off-by: Nir Soffer <nsoffer@redhat.com>