lib: jsonrpc: use periodic warning on stuck worker
In change 54b79818e17e44ba1acf7d47e79a7e7946b50b4f we added the option
to check and warn in the logs about stuck executor workers instead
of discarding them.

This patch uses this option for the jsonrpc workers, which were
among the hardest to troubleshoot in case of exausted executor pools.

Change-Id: I62db37cc411307322564edc467abcdff8278fbee
Bug-Url: https://bugzilla.redhat.com/1428415
Backport-To: 4.1
Signed-off-by: Francesco Romani <fromani@redhat.com>