virt: Don't patch *VolumePath in test_change_cd_failures
There is no need to monkey patch those methods to make the test pass.
Actually patching prepareVolumePath leads to an XML error that is raised and
silently swallowed.  In the result, we test for a failure, but a wrong failure:
Not the libvirt one, but the XML formatting failure.

We additionally need this fix for a followup patch where XML formatting is
taken out of the corresponding try...except block.

Backport-To: 4.1
Bug-Url: https://bugzilla.redhat.com/1428514
Change-Id: I303c1f84653b1a76008405904a22441873ddc224
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>