virt: drop unused guests_gateway_ip
guests_gateway_ip is a prehistoric configurable, that has been
completely replaced by display network.

If it is ever used, it renderes a VM non-migratible (as the graphics
address on one host should not exist on the destination host, too).

This patch drops something that should have been removed many years ago,
with the introduction of libvirtvm.py.

Change-Id: Ia2e08388ffab439cb8985d4f220a410b4f6b6c90
Signed-off-by: Dan Kenigsberg <danken@redhat.com>