image: Allocate initial size for converted RAW volume.
Allocate the initial size of a QCOW2 volume on block domain which was converted
from RAW volume.

This code will be used on create template from VM, create VM from template (on
DC 4.0), import a VM/Template from export domain, and also many trivial use
cases that includes copy or move.

Up until this fix was introduced, VDSM used to pre-allocated all the virtual
size of a RAW volume once it was converted to QCOW2.

The effect of this fix will be that there will be much less space to allocate
since it will be dependant on the actual size of the volume.

Currently this fix will only be supported for source RAW sparse volumes hosted
on NFS v4.2.  Lower NFS versions than 4.2 will be pre-allocated with the
virtual size of the volume.  Also source volumes which are pre-allocated will
be converted to pre-allocated volumes as well.

For example a RAW disk created on V4.2 NFS server and converted to COW volume:
The source volume is RAW with actual size of 1.23GB and virtual size of 10GB

Converting this image to qcow2 allocates now 2.25GB.
The image itself requires 1.232GB, and we allocate extra chunk (1G by default),
to make room for growing before the image must be extended.

Change-Id: I6a0aacc8c66327b2c2d6d0a024746ca41fac2fe4
Bug-Url: https://bugzilla.redhat.com/1358717
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>