api: adding context for api calls
We want to understand how specific requests pass through the system and
we are using flow id for that. We get it from a client using custom
header and it is passed around in vdsm using threadlocal.

The context that we pass consists of:
- flow_id - uniquely identifies flow
- client_host - host ip address from where specific verb was called
- client_port - which port was used to call the verb

When a client do not provide flow_id we use None to indicate it.

Bug-Url: https://bugzilla.redhat.com/1112120
Change-Id: I8e73edc929e6d3e55d3f33a51229180ba35c52ac
Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>