virt: Switch to persistent domains
We have been using transient domains for running VM so far.  They disappear
when their VMs stop running at the given host.  Detailed information about a
destroyed VM is still retained in its recovery file, until Engine asks for
releasing it.

We would like to stop using Vdsm specific recovery files and use libvirt domain
XML as the complete and ultimate source of information about a VM instead.  In
order to be able to access the domain data before the VM is started or after
it's stopped, similarly to the recovery files, we need persistent domains that
don't disappear immediately.

This patch switches from transient to persistent libvirt domains.  Creation and
removal of the domains is handled similarly to recovery files.  The only extra
feature is that we must remove the domain, if it exists, before a VM is
defined.

Change-Id: I7fa319d7e208b5c8e06f7e31b4bc47e5cc1b4d67
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>