virt: Store VM start time to metadata
We want to stop using recovery files in the future.  In order to discard
recovery files, we must store all relevant information to domain XML.  If some
information is not available from domain XML or libvirt functions, we must
store it to domain XML metadata.

This patch introduces the needed functionality:

- It defines a method to store needed information about a VM to metadata.
- It calls the method when the VM is started.
- VM start time is stored to metadata as a particular piece of needed information
  currently not available elsewhere.  Other pieces of information will be added
  in followup patches.

We plan to update metadata only when a VM is started and at (rare, if any)
occasions when information in metadata changes.  This a significant difference
to updating recovery files, which must be updated at each status or device
change.  With information stored in domain XML and VM status determined from
libvirt (once https://bugzilla.redhat.com/1441563 is fixed) such frequent
updates are no longer necessary.

Change-Id: Ifd27f0144bdfb21f47126b5c6d696a68166e4056
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>