net tests: removing the remainder of *CompatibilityMultipleNetsSameNic
After having portions of the test moved to the new functional tests,
what's left is redundant. The remainder of the test initially creates 2
networks:

1) 'test-network4' - 'another-nic', vlan: 100
2) 'test-network5' - 'nic', vlan: 200

There's no value in checking whether these networks were created
successfully as the use case is already covered(2 different networks,
different nics, different vlans)

Afterwards supposedly 2 networks are created, but it is not so as the
networks have the same name and are part of a single transaction
(the second network simply overwrites the first network as they have
the same dictionary key).

This leads to a single new network:

3) 'test-network6' - 'yet-another-nic', vlan: 300

Another network with a different name on a different nic with a
different vlan tag.

Change-Id: I357fadd683f9fe3f890eb5645f845b8538dc59b5
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>