monitor: Fix pylint error when invoking callabck
Pylint was complaining about non-callable cycleCallback. The callback
was initialized to None, and the code was checking if cycleCallback is
not None before invoking it, but pylint is not smart to enough to
understand this.

Fixed by initializing to _NULL_CALLBACK function. Now we don't need the
None check and we can invoke the callback on every iteration. This is a
little bit more expensive compared with the None check, but we call this
once in 10 seconds so it does not matter.

Change-Id: I8f7ac8e842d415b7fe457875529abcae35ddcaec
Signed-off-by: Nir Soffer <nsoffer@redhat.com>