pylint: Remove useless code from IscsiInterface.__getattr__
The original code was trying to access object.__getattr__ for getting
existing attributes in __getattr__. This is wrong because object does
not have __getattr__, but the code was safe since __getattr__ is invoked
only for missing attributes, so the wrong code never run. Remove the
useless code.

For more info see:
https://docs.python.org/2/reference/datamodel.html#object.__getattr__

Change-Id: Ie9f12797df2ba7abe87ce5ecb733caeacc535233
Reported-by: pylint
Signed-off-by: Nir Soffer <nsoffer@redhat.com>