vm: blockIoTune: add and use cache
Getting and setting the block I/O tune values need to go
deep down the QEMU monitor. Thus, this call is costly and risky,
and we should avoid it as much as we can for obvious benefit purposes.

MOM and Engine needs this data, but both pass through Vdsm
to get and set the values.

Thus, it makes sense to add a caching layer for this data inside Vdsm.
Doing so we both increase the system performance reducing the libvirt
load, and by relieving stress on libvirt side we reduce the risk
of thread pool exhaustion.

Bug-Url: https://bugzilla.redhat.com/1443654
Backport-To: 4.1
Change-Id: Ie50fc8a9395eab60a8b97b998555df1146af5b1b
Signed-off-by: Francesco Romani <fromani@redhat.com>