pylint: Remove broken and unneeded code
oop.os.rename() was trying to handle the case when the destination
directory exists and is not empty. In this case, the code tried to
delete the destination directory, and rename the source directory.
This code was broken, failing with:

    AttributeError: '_IOProcessOs' object has no attribute 'listdir'

The broken code was introduced in commit 576428e876c2 (ioprocess
implementation), and is available since 4.16.0.

The only effected flow is Image.delete, when we rename the image
directory to remove_me_<uuid>. In this flow, we do not expect that the
destination directory to exist.

This patch removes the unneeded functionality, restoring POSIX
semantics.

Change-Id: I352e8643a47f517c879d2085bced6847e0372b30
Bug-Url: https://bugzilla.redhat.com/1465836
Reported-by: pylint
Signed-off-by: Nir Soffer <nsoffer@redhat.com>