pylint: Simplify pylint unfriendly code
Pylint can not check properly code based on a type check. Replace the
type check and normalizing with simple search.

This also move argument normalization to the top of the method, so force
is always a boolean inside this method. Add TODO to remove this
normalization later, we don't want to support this.

Change-Id: I64c2b1855fee01a6aaaa0ffbc9506267d190bf2d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>