pylint: Unify removeMetadata interface
removeMetadata() is defined differently in BlockVolumeManifest (accepts
a metaId argument) and FileVolumeManifest (doesn't accept any argument)
while it is not defined in VolumeManifest.  This works in runtime since
this method is called only from the concrete class with the correct
number of arguments, but it makes pylint unhappy and makes the code
harder to understand.

Add the missing abstract method in VolumeManifest, and an optional
metaId argument to FileVolumeManifest so the interface is similar.

Change-Id: Ie4b24bc5da5ca2dcf60151cf062afc5f356e6d5f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>