network: fail network validation when adding dpdk device
Adding dpdk nic is only supported in case of an ovs switch,
and only on an ovs bridge.
This patch prevents the user from using dpdk nics as part
of a legacy network or as slaves of a linux bond.
It also prevents adding a dpdk device as a slave of ovs bond.

Change-Id: If7394428f14a6b0db2f3aaee798daa767ab1aae6
Signed-off-by: Irit Goihman <igoihman@redhat.com>