vmpowerdown: implement base class interface
The VmPowerdown class was not implementing fully its interface,
demanding to its derived class to fill the blanks. This made
pylint unhappy and the code a bit harder to follow.

In this patch we complete the VmPowerdown interface, so the derived
class need now to override the attributes.
This makes pylint happy and the code a bit easier to read.

Change-Id: I1a5f7880c5f3e8cc37c979294a501632d8dcbe8d
Signed-off-by: Francesco Romani <fromani@redhat.com>