net: Centralize bond mode 'numerization'
When handling the bond mode, at several location the numeric value is
required, therefore the conversion from a name value to a numeric value
is performed.

As this task is currently perform at several locations outside the bond
modules, this patch centralizes it under one of bond modules.

This change originated from the need to add another such numerization of
the bond mode value.

Change-Id: I3e40fe1b539a9844a3bb40266d99a87ff4d4f628
Signed-off-by: Edward Haas <edwardh@redhat.com>