vm: Detect when a block job is ready
When a block job is ready, blockJobInfo returns equal "cur" and "end"
values, but equal values do not mean that a block job is ready. To get
the ready state, we can use the disk's mirror element:

    <mirror type='file' job='active-commit' ready='yes'>
      <format type='raw'/>
      <source file='/rhev/data-center/.../168134ea-a173-448c-b1b9-087abac87ed8'/>
    </mirror>

The ready flag was not reliable before libvirt-1.2.18. Since we require
libvirt >= 2.0 on all platform it should be safe to use now.

Change-Id: Ic0b3da576c3eca56b3754e74ae4cd9c986b0fcec
Bug-Url: https://bugzilla.redhat.com/1442266
Bug-Url: https://bugzilla.redhat.com/1438850
Signed-off-by: Nir Soffer <nsoffer@redhat.com>