ssl: change default library
We would like to change default ssl library being used from m2crypto to
ssl. The value that we saw when introducing m2crypto was no longer
there when we started to maintain the connection. We do not need to
maintain two code bases using different ssl libraries.

This patch would enable us to remove m2crypto in the future.


Change-Id: Ib6c64b0781cee9d97829e955490e0608e4208933
Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>