virt: Initialize DomainDescriptor early
We are going to retrieve various information about the VM from its domain
XML (instead of vm.conf), even before the VM is actually started.  So we should
initialize the domain descriptor from available data as soon as possible.

The initialized domain will still be incomplete and must be amended later as
devices are added in Vm._run.

Change-Id: Ie5001aed730de2988d436deec7fdf65588808726
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>