metadata: namespace support in _find_device
Will be used by *from_xml functions.
Non change in existing behaviour.

We need to explicitely support namespaces because
we access the same XML data from two sources.
Once the domain is defined, libvirt takes care of the
namespace handling through its API, so the XML we work with
is clean and simple.

In the creation flow, however, we must deal with the XML
ourselves and take care of the namespaces.

The metadata code needs to support both those flows, hence
the need for explicit, optional namespace support.

Change-Id: If68cf2c625ae44144093816890fe07f07082ab09
Signed-off-by: Francesco Romani <fromani@redhat.com>