vm: update vm properties from metadata
Since we are going to get all the VM parameters from the
client-provided domain XML, we now need to get the custom properties from
the domain metadata.

Furthermore, we also need to ensure we always haver the 'devices'
key in the vm.conf dict, even though we no longer use it, to avoid
to pepper the Vm class code with `if`s.

Change-Id: Ide8b89f8e7d68b8d56b0dffcd8844cedc88fac0a
Signed-off-by: Francesco Romani <fromani@redhat.com>