virt: host: refactor the VM memory usage reporting
This patch rewrites the reporting of the memory usage of the VMs,
eliminating the need for the ugly memCommit and vm.conf['pid']
attributes.

We use one new call now, and we rely on libvirt to get the memory usage
info for the VMs (see
https://libvirt.org/html/libvirt-libvirt-domain.html#virDomainMemoryStats).

No expected change in behaviour, only internal refactoring.

Change-Id: I5e1414b6dcc2f3312a69dc354a9d59d95354a747
Signed-off-by: Francesco Romani <fromani@redhat.com>