image: prepare volume before calculating size
In image.copyCollapsed, there is a need to calculate the destination volume
blocks needed to be allocated before creating it.

As part of the calculation, if the source volume is raw, qemu-img info
is called to read the actual size in use.

In case that the source domain is a block domain, the volume need to
activated before it can be read.

In this patch, the 'prepare' of the source volume is moved to be
executed before the call to 'calculate_vol_alloc'
The volume will be teardowned as before in the 'finally' of the method.

Change-Id: Ifdad6b850b25f06d3333a6fd810d91f9d8c0d183
Bug-Url: https://bugzilla.redhat.com/1448606
Signed-off-by: Fred Rolland <frolland@redhat.com>