blockSD: Allow using badly restored storage domains
Block Storage domain is created with 2 metadata areas on all PVs, but
all the metadata areas are disabled excpet the first PV. This is
typically the save PV where the metadata LV is located. Since 4.1 we are
reporting the metadata PV in StorageDomain.getInfo. If we cannot find
the metadata device, we will fail with UnexpectedVolumeGroupMetadata
exception, and activating the storage domain will fail.

In the error, we can see the PVs with incorrect configuration
(irrelevant attributes stripped):

    PV(name='/dev/mapper/360014054ae1fdf75b074e488e9d803bd',
       ...
       mda_count='1',
       mda_used_count='1'),
    PV(name='/dev/mapper/360014053f404fa44d844d9198cfee437',
       ...
       mda_count='1',
       mda_used_count='1')

We can see that both PVs have one used metadata area, instead of 2 used
metadata areas in one (the metadata PV), and 2 unused metadata areas in
the other (regular PV).

Turns out that there are lot of users that somehow got their VG metadata
deleted, and unfortunately used incorrect instructions for restoring the
VG metadata.  This process did not use the required "--metadatacopies 2"
option, and the "--metadataignore y" option on all the PVs but the
first. This problem is too common so we cannot fail
StorageDomain.getInfo in this case.

This patch changes the behavior to log a warning, and return the storage
domain info without the "vgMetadataDevice" key. The user will not be
able to remove a PV from such storage domain.

Storage domain with this configuration is not supported, and must be
fixed by recreating the LVM metadata correctly.

Change-Id: Ie04922f951423d032c81a0212372fb248c8d7af7
Bug-Url: https://bugzilla.redhat.com/1446492
Signed-off-by: Nir Soffer <nsoffer@redhat.com>