tests: testlib: make read_data pick the right cwd
testlib.read_data wants to make easy for a testcase
to read some static data.
To do so, it does some smart magic to automatically
discover the Current Working Directory (CWD) and pick
the specified file from the $CWD/data subdirectory.

The current implementation, however, detects the
CWD of the *defining* module (tests/testlib.py),
not the one of the *calling* module.

This either requires to duplicate the per-vertical hierarchy like

tests/
 +-data
 |  +- virttests/
 |  +- network/
 |  +- storage/
 +- virttests/
 +- network/
 +- storage/

or makes the helper useless.
This patch adds a bit more magic to read_data to make it
detect the CWD of the calling module, so we can have this
hierarchy instead, fully separated by vertical:
tests/
 +- virttests/
     +-data/
 +- network/
     +-data/
 +- storage/
     +-data/

Change-Id: Ia19a025301b24653f25ec3ce3a7ee1c1e02c407a
Signed-off-by: Francesco Romani <fromani@redhat.com>