dockerfile: Add missing glusterfs-api package
This package is required for vdsm.gluster.gfapi module. Importing this
module during imports tests fail on Fedora 25 becuase libgfapi is not
installed in this image. For some reason, this library is included in
the Fedora 24 and Centos 7 image.

Add the library to all dockerfiles to make sure the library exists even
if the base image is modified in the future.

Change-Id: Ieaf94a2d3a94cb6693893db8de70063ad1503510
Signed-off-by: Nir Soffer <nsoffer@redhat.com>