tests: Add tests for logged methods
Testing vars.context is tricky, since it exists only if the thread
owning the thread local data has set vars.context.

This patch adds a helper to run a function in a thread, setting
vars.context, and use the helper to test that @logged methods work when
a context is set.

The logged tests should test the logs before and after the call, but the
way @logged is defined makes this too ugly. We will add the missing
asserts later.

Change-Id: If17a9401c0666987de5b804e79206fe111472ca3
Bug-Url: https://bugzilla.redhat.com/1112120
Signed-off-by: Nir Soffer <nsoffer@redhat.com>