net test: Cleanup colliding test ip rules
During the CI test runs, stale test ip rules have been detected on
the nodes, causing test failures.
Failty tests (or production code) are suspected to run on the nodes,
causing their network stack to become dirty.

This patch is introducing a package teardown handler that checks if
there are any test ip rules left on the system and attempts to remove
them.
An exception (StailIPRulesError) is raised in case such ip rules are
detected to inform of this problem.

Note: This is a test level attempt to make sure bugs do not leave the
node dirty. However, it will not cover all corner cases (like a crash of
the run before reaching the cleanup section), therefore infrastructure
solutions are still needed (netns, containers, VM) and should be
considered in order to protect the nodes.

Change-Id: I461ad5697525cfaa2e3c6cfa66269692283903ce
Signed-off-by: Edward Haas <edwardh@redhat.com>