vm: Remove unneeded return value handling
Some libvirt calls are documented to return 0 on success and -1 on
error. However, when using the python bindings, these calls always
return 0, or raise an exception. Therefore, there is no need to check
the return value, only handle the exception.

This patch removes unneeded return value check when starting live merge.
This is the libvirt binding implementation:

    ret = libvirtmod.virDomainBlockCommit(self._o, disk, base, top, bandwidth, flags)
    if ret == -1: raise libvirtError ('virDomainBlockCommit() failed', dom=self)
    return ret

Change-Id: I2f54a5c23bb0ac13d5c7f158bec8f6938046b10f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>