logUtils: Move logUtils module to vdsm.common
logUtils has been relocated to common, to make it clearer
that it is being used by multiple subsystems, and verifying that it
has no dependencies on non-common vdsm code.

The module has been renamed to logutils.

This patch also fixes lib/vdsm/common/api.py which depended on an
external (non common) module.

A follow up patch is planned to add traceback to this module.

In order to support upgrades of systems that have manually edited the
logger conf file, a small proxy is left in the current logUtils module,
pointing to the new location under common.

Change-Id: Ic642205e3d5515acfeb607bb5dcad24f699d286f
Signed-off-by: Edward Haas <edwardh@redhat.com>