cold-merge: Minimize illegal chain during cold merge
Before this patch, base volume was set to ILLEGAL at the prepareMerge
step. That decision was too conservative as at the prepareMerge step the
volume size is updated, and at the merge step, the data is copied. The
critical point where setting base volume to ILLEGAL, is when updating
qemu metadata in finalizeMerge step.

We don't change base volume legality anymore, because it remains valid
during the whole merge process. However, we set top volume to ILLEGAL
in the very short window when updating qemu and vdsm metadata.

This patch also enhances recovery from failures during cold merge.
Previously, if cold merge failed at any step, the base volume was
ILLEGAL and manual recovery (fixing the metadata) required. With this
change, the chances that the top volume is ILLEGAL are really low, which
enables the user to recover from failures by retrying cold merge.

Change-Id: I86b76ab4e62de6e75f20cfdb6ac471e73bbcecc6
Bug-Url: https://bugzilla.redhat.com/1430358
Signed-off-by: Ala Hino <ahino@redhat.com>