storage: Add index support for blockcommit
libvirt virDomainBlockCommit call is used in merge flow and requires
two arguments - base file and top file, commiting changes that were
made to temporary top-level files within a disk image backing file chain
into a lower-level base file.

Those base and top arguments could be either paths to files within
the backing chain, or the device target shorthand, followed by an index
to the backing chain enclosed in square brackets.

For example it could be either '/path/to/file' or 'vda[1]'

Before that patch vdsm used first approach and supplied
virDomainBlockCommit call with path to files.

Unfortunately it appeared, that using indexes is a recomended approach,
due to qemu limitations:

"the URI itself can't be really identified by libvirt in the API entry
point, since that depends on the qemu impl itself, since that API does
not take an XML which is the frontend description of the disk.
But with using indexes in the backing chain, libvirt will know how
drive ("vda") is implemented internally."

We actually hit that limitation with network disk types, as per
https://bugzilla.redhat.com/1397870

As network disk types require using indexes in block commit calls and
using indexes is a recommended way to supply base/top arguments of that
call, this patch switches merge flow for all disk types to indexes.

Change-Id: I6c9a1017be7ee21af8796ebf6abff35c13fa9c99
Signed-off-by: Sahina Bose <sabose@redhat.com>
Signed-off-by: Denis Chaplygin <dchaplyg@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1022961