storage: Don't swallow domain lock init failures
When creating a new storage domain, we were swallowing errors when
initializing the SPM lease failed.  This just delays problems because an
incomplete domain is created and subsequent flows (eg. attach SD) will
fail due to a missing Sanlock lockspace.  All new domains must have a
functional cluster lock so stop ignoring failures.

Refactor the test classes so that tests meant to be run on file and
block domains can be implemented just once in a shared ancestor class.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1386830
Change-Id: I9bc8c1db1804615c1d270a641c34e6eb1b50d5a2
Signed-off-by: Adam Litke <alitke@redhat.com>