net: Introduce a simplified cmd exec implementation
The network package is frequently utilizing the execCmd function to
execute shell based commands.
This patch introduces a simplified version of execCmd, focused on the
limited needs of the network package.

The new cmd module exposes a blocking exec function (exec_sync) and a
systemd exec function.

In the following patches, the users of execCmd will get migrated to the
new cmd module implementations, reducing the dependency on the generic
VDSM libs.

Change-Id: I37380d51b648bd64792b6c26419dd33c96392ddd
Signed-off-by: Edward Haas <edwardh@redhat.com>