net test: Refactor NM tests and its test lib
The patch includes NM tests setup and teardown functionality, making
sure the resources are cleaned from the VDSM configuration.

The dummy nics that have been used as slaves to the NM bond are now
created at the test level (and not in the test lib) to allow more
control and reduce the complexity in the test lib.

Change-Id: I6dd262c40607088613ab9be4f1911e11f1926e47
Signed-off-by: Edward Haas <edwardh@redhat.com>