net test: Use a single slave in the NM vlan-bond-slave sync test
The test_add_net_based_on_existing_vlan_bond_nm_setup test has randomly
failed due to the slaves mac address swap (the bond mac is assigned
based on the first slave added).

Until the NM and ifcfg/initscripts use a consitent order to load the
slaves, the test will use a single slave.

Note that the test itself was focused on the vlan iface mac being
identical to the bond mac and not on the consistency of the slave load
order.

Change-Id: I65ef31973198914e22d0d8da08b6a81e386ca533
Signed-off-by: Edward Haas <edwardh@redhat.com>