net: Remove iproute2 configurator
ifcfg and iproute2 are the existing available net configurators but
iproute2 has not been maintained (or tested) for some time now.
It is also not in use by any module.

As we plan to replace the configurators as defined today, there is no
reason to keep iproute2 in the code base.

Change-Id: I8a5de53fa64a74fffbf97ca9655a7303f17ae25c
Signed-off-by: Edward Haas <edwardh@redhat.com>