tests: simplify _check_roundtrip
_check_roundtrip used to return the instance it created after the
check succesfully completed. That allowed to piggyback more tests after
the main one.

Now we remove this ability for the sake of cleanliness - and to further
improve the roundtrip tests: in a future patch
(Ie2cdfde781d929c3575fe170ffd2e1b35e850e70), we will need to call
setup/teardown around getXML() in _check_roundtrip, so it makes no
sense anymore to return the object we used for the check: after
teardown, it will become useless.

The two tests that relied on the previous behaviour of
_check_roundtrip are being split off and preserved.

Change-Id: I5971902559b2845ec4c18d890f96d04baac1b477
Signed-off-by: Francesco Romani <fromani@redhat.com>