tests: call device setup/teardown
The proper device protocol requires that the Device instances
are completely initialized before to call any real method on that,
otherwise we end up in undefined behaviour (method may or may not work
depending on device parameters and the method being called).

We accomplish that doing

obj = Device(*args, **kwargs); obj.setup()

The current test didn't do that, because core devices don't need to do
anything meaningful in their setup().
The only exception was the Rng device, but here the bug sneaked in
because of the test parameters, which do not mandate the instance to
do further actions in setup()

This patch ensures the right protocol is followed in _check_roundtrip,
making the test code ready to accomodate the complex devices (Graphics,
Interface), which will need to do complex operations in
setup()/teardown().

Change-Id: Ie2cdfde781d929c3575fe170ffd2e1b35e850e70
Signed-off-by: Francesco Romani <fromani@redhat.com>