net: Route removal should not ignore the table it is in
When querying the list of ip routes in a specific table,
the returned list in iproute2 implementation does not include the
table ID and therefore it was wrongly assumed as 'none'/any.

The table is returned in the report only if the table is 'all', which
reports all routes from all tables.
The ip.route/iproute2 driver has been updated to include the correct
table for each route data object.

Change-Id: I628d5db06e567a13056dbf83ad7b4d17d79b4ac6
Bug-Url: https://bugzilla.redhat.com/1452185
Signed-off-by: Edward Haas <edwardh@redhat.com>