virt: Improve migration check
We need to know in domain XML based recovery and when handling external VMs in
the future whether a VM is migrating to another host.  Current
migration.ongoing utility checks just for a running job regardless of its
operation, since it assumes that we are migrating.  This patch enhances the
function to check also for the job operation kind.

The functionality necessary for the improved check is present only in recent
libvirt, not yet universally available (see
https://bugzilla.redhat.com/1441563).  So we can't rely on its presence now and
we must not fail with current libvirt versions.  The check can't distinguish
between job operations with those libvirt versions but otherwise it's not worse
than the current state.  It will get fixed automatically once new libvirt is in
use.

Change-Id: I5c10caf22f197facae857138740dbc3122fd9b54
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>