virt: Always check migration status on recovery
When we recover a VM we check its migration status only if migration is
indicated in the VM's recovery file.  But we want to stop using recovery files
and we also want to support external VMs.  In both the cases there is no
special hint whether the VM is migrating, so we must always assume that
possibility.  For that reason we always check for a possibly running migration
in recovery.

Change-Id: I6aed87092792674e6c47fd6dd7524c1eb3bf058f
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>