network: decode command output to native string
We should keep strings native. With this patch we decode command
output from binary to unicode (in case Python 3 is used) directly
in network.cmd.exec_sync. Thanks to that we handle all strings
natively and we don't end up with errors trying to split binary
with unicode.

This patch fixes Python 3 problems where we handled command output
as a str although it was bytes.

This problem lead into failed tests with Python 3 running. Those
tests could be removed from Python 3 blacklist as soon as we fix
other Python 3 problems.

Change-Id: I39ed13161487cff4f531f513a25a1c74787cf2df
Signed-off-by: Petr Horacek <phoracek@redhat.com>