clusterlock: Wait until host id is acquired
acquireHostId() is typically called with async=True from the domain
monitor. Then the domain monitor invokes hasHostId() every monitoring
cycle, until the operation completes. Vdsm reports the acquire status of
a domain to engine, but unfortunately engine is ignoring the acquired
status, and try to start storage jobs before the host id is acquired.
This leads to failure to acquire a volume lease if a job was started
before the asynchronous host id acquire operation was completed.

This patch changes SANLock.acquire() to wait until a host id is acquired
before trying to acquire a lease. If engine choose a bad host that did
not acquire the host id yet, the job will be delayed up to 30 seconds in
normal conditions when trying to acquire the volume lease.

If sanlock cannot acquire the host id, a job may be delayed up to 180
seconds waiting for host id acquire completion.

Flows that acquire a temporary host id before acquiring a lease should
not be effected by this change.

I checked another approach, trying to acquire the host id if it was not
acquired yet. This creates an ownership issue with the domain monitor,
managing the host id acquire life cycle, so it was rejected.

Change-Id: I110e05f13549b27fef1700d97a00c0efb71a3bc0
Bug-Url: https://bugzilla.redhat.com/1422508
Signed-off-by: Nir Soffer <nsoffer@redhat.com>