clusterlock: Move logs out of the locked region
Move the log before performing an operation and after the operation
completes out of the locked region. The main reason for this change is
to make it easier to debug when multiple threads try to call the same
code protected with the same lock.

This change also modify the log level of the completion log to INFO, to
make it easier to understand when the operation was finished. Without
the completion logs, you have to guess the completion by searching for
unrelated logs in in in the same flow.

Change-Id: Ia462adc918b0dc7f708b831b34142103977f11e7
Bug-Url: https://bugzilla.redhat.com/1422508
Signed-off-by: Nir Soffer <nsoffer@redhat.com>