spec: Clean up lvm2 requirement
The epoch in lvm2 RPMs on EL distros represents the major version
(see [1]). As we only support EL7 (and will for the next couple of
years), it's safe to drop the epoch requirement from the spec file and
simplify it.
Moreover, if lvm2 continues using monotonically increasing version
numbers, there would be no reason to use the epoch even when we do
support EL7 and EL8, unless, perhaps, there's a specific fix
backported to EL7 we'd like to require - but that bridge can be
crossed when we get to it.

[1] https://git.centos.org/blob/rpms!lvm2.git/191583678466169c13b312069ea9d83d1e958d39/SPECS!lvm2.spec#L662

Change-Id: I411d43dc57dfc37702a0125c24b6281a480cf6de
Signed-off-by: Allon Mureinik <amureini@redhat.com>