spec: Handle VDSM services restart on upgrade to posttrans
VDSM services restart has been handled upon upgrade at the postun stage,
which is triggering the scriptlets of the source/old version.

In order to support upgrades with new logic, the scriptlets should run
based on the new version.
Therefore, the handling has been moved to the posttrans step, which
belongs and runs as part of the new version at the end of the
transaction.

Side effect: Upgrades from versions that predates 4.2 will have the
service restart occuring twice, once at the old version postun step and
once in the new posttrans step.

Change-Id: I16793ff6ca1fcf596f57b459fa4ed348a879b104
Bug-Url: https://bugzilla.redhat.com/1457889
Signed-off-by: Edward Haas <edwardh@redhat.com>