virt: Set correct diskType for snapshot target drive
Snapshotting code now checks for correct protocol of
base and target drives in case of NETWORK disk type,
but setting correct protocol in prepareVolumePath
requires diskType to be correctly specified before the call.
As we do not support snapshots on different disks, it
can be safely copied from base drive to the target drive.

Change-Id: I36fe89a0c40ac6c81c38fdd12fe9e945ac5aa901
Signed-off-by: Denis Chaplygin <dchaplyg@redhat.com>