virt: vmstats: never raise if stat is missing.
We used to try to distinguish if the lack of stat reported by libvirt
is a serious issue or could be ignored. There are a lot of flows
and circumstances to consider, making this a daunting task.

Ultimately, all this effort is misplaced, because even if we can
reliably detect that the lack of one expected stat is a bug, we can
do nothing besides record this event in the log and carry on.
The only noteworthy exception to this is if we somwhow ask for a key
that can't exist; for this reason, better not to completely silence
the failure.

To handle those cases, one concise log is enough, no need to bubble
up the exception and to spit scary stacktraces in the logs.

Change-Id: Ib7f19fe26151fbc4d9fe6ff9c83ca3db77ff6e25
Bug-Url: https://bugzilla.redhat.com/1458901
Signed-off-by: Francesco Romani <fromani@redhat.com>