virt: store balloon target in Balloon dev object
We call 'balloon target' the current amount of memory a Vm could use.
The modelling of this property is already debatable: we store it
as Balloon device attribute ('target', unsurprisingly), yet we use
domain methods to set it (virDomainSetMemory).

The bugger issue is that up until now we stored this setting as
device configuration item, thus in Vm.conf['devices'], and never
propagated it to the device object proper.

This will cause issues when we will migrate to engine XML, because
the vm.conf will be deprecated.

In this patch we streamline how we use the 'target' property, as
follows:
1. we make it a full blown Balloon device attribute
2. we use it as authoritative source - no more the vm.conf['devices']
   counterpart
3. we ensure we initialize it from the vm.conf, to handle recvoery
   gracefully.

No expected changes in external behaviour

Change-Id: I58e63c5302f54e91a717afb5126d62feb0093233
Signed-off-by: Francesco Romani <fromani@redhat.com>